requirements.txt
asgiref
async-timeout
certifi
cffi
charset-normalizer
cryptography
Django==3.2.20
django-cors-headers
django-mongodb-engine
django-redis
djangorestframework
djangotoolbox
djongo
dnspython
idna
pycparser
PyJWT
pymongo==3.12.1
python-dotenv
pytz
redis
requests
six
sqlparse
jsonfield
typing_extensions
urllib3